Fix EnableAmendment tx scanning process and add tests (#203)
## High Level Overview of Change

<!--
Please include a summary/list of the changes.
If too broad, please consider splitting into multiple PRs.
If a relevant Asana task, please link it here.
-->

The amendment's ETA when reached majority or enabled was not captured by
the VHS at the moment. This PR attempts to fix the issue by:

- When an `EnableAmendment` transaction is found, use ledger's
`close_time` instead of transaction's date field in the `ledger`
response, which does not exist (This will also simplify the code as
extra `tx` request does not need to be sent to capture the date).
- Remove `date` from ledger response types transaction.
- Add test for `handleWsMessageLedgerEnableAmendments` function which
scans for `EnableAmendment` transaction at Flag + 1 ledger.
- Add more logs to the amendment scanning process.

### Type of Change

<!--
Please check relevant options, delete irrelevant ones.
-->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (non-breaking change that only restructures code)
- [x] Tests (You added tests for code that already exists, or your new
feature included in this PR)
- [ ] Documentation Updates
- [ ] Release

## Test Plan

<!--
Please describe the tests that you ran to verify your changes and
provide instructions so that others can reproduce.
-->
Leave on staging for couple days to ensure all flag+1 ledgers are
scanned.