feat: Use feature for amendment name (#291)
## High Level Overview of Change

<!--
Please include a summary/list of the changes.
If too broad, please consider splitting into multiple PRs.
If a relevant Asana task, please link it here.
-->
Resolve: https://github.com/ripple/validator-history-service/issues/278

### Context of Change

<!--
Please include the context of a change.
If a bug fix, when was the bug introduced? What was the behavior?
If a new feature, why was this architecture chosen? What were the
alternatives?
If a refactor, how is this better than the previous implementation?

If there is a design document for this feature, please link it here.
-->

Missing new amendments due to the change in rippled code structure to
parse name:

![Screenshot 2024-12-19 at 11 32
43 AM](https://github.com/user-attachments/assets/0da97dff-9d6d-4840-b8cf-6b44ba628053)

### Type of Change

<!--
Please check relevant options, delete irrelevant ones.
-->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (non-breaking change that only restructures code)
- [ ] Tests (You added tests for code that already exists, or your new
feature included in this PR)
- [ ] Documentation Updates
- [ ] Release

## Before / After

<!--
If just refactoring / back-end changes, this can be just an in-English
description of the change at a technical level.
If a UI change, screenshots should be included.
-->
Data should show up:

![Screenshot 2024-12-19 at 11 58
52 AM](https://github.com/user-attachments/assets/5198638f-4c69-4158-9a31-fa004aab78c4)

<!--
## Future Tasks
For future tasks related to PR.
-->
VHS is lacking tests at the moment. There will be a separate ticket to
add comprehensive tests throughout the code