fix: add unl blob v2 (#304)
## High Level Overview of Change

<!--
Please include a summary/list of the changes.
If too broad, please consider splitting into multiple PRs.
If a relevant Asana task, please link it here.
-->

UNL v2 was deployed to devnet in the most recent reset. This introduces
a field called `blobs_v2` which is an array containing multiple blobs.
The VHS needs to be fixed to capture this change.

This fix has been deployed to staging:
https://stg.validator.history.xpring.dev/v1/network/validators/dev
And also could be checked at:
https://dev.dev.explorer.xpring.dev/network/validators

### Type of Change

<!--
Please check relevant options, delete irrelevant ones.
-->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (non-breaking change that only restructures code)
- [ ] Tests (You added tests for code that already exists, or your new
feature included in this PR)
- [ ] Documentation Updates
- [ ] Release