Exponential backoff before WebSocket connection retry (#325)
## High Level Overview of Change

This change is being introduced to fix hourly agreement not being
updated from xahau-main network.

1. Added exponential backoff before WebSocket connection retries, so
that the validator server is not swamped with connection requests and
hopefully, our server's ip address is not blocked from any further
requests.

2. Removed `connectionsInitialized` check from websockets `close`
callback before retrying. Since, as per the current design of the code,
if all the connections are not initialized and one of the connection
(lets say wss://xahau.network/) closes , it won't be retried until next
hour.


### Context of Change

To fix connection issue specifically for `wss://xahau.network/`

### Type of Change

<!--
Please check relevant options, delete irrelevant ones.
-->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (non-breaking change that only restructures code)
- [ ] Tests (You added tests for code that already exists, or your new
feature included in this PR)
- [ ] Documentation Updates
- [ ] Release

## Before / After

<!--
If just refactoring / back-end changes, this can be just an in-English
description of the change at a technical level.
If a UI change, screenshots should be included.
-->

## Test Plan

This issue occurs only in production. So, we will deploy this branch on
staging and see if the intended fix do not cause any issue.

<!--
## Future Tasks
For future tasks related to PR.
-->