add `ledger_hash` values in `/network/validators` endpoint (#331)
## High Level Overview of Change

This PR aims to fix
https://github.com/ripple/validator-history-service/issues/67. This is
an old issue, I'm not sure if its still required, I did this for fun.

<!--
Please include a summary/list of the changes.
If too broad, please consider splitting into multiple PRs.
If a relevant Asana task, please link it here.
-->

### Context of Change

The `validators` table already contains the `ledger_hash` column in its
schema. This PR merely updates the external API endpoint at
`/network/validators`.

This PR does not introduce any breaking changes. It adds a new field in
one of the API responses.

<!--
Please include the context of a change.
If a bug fix, when was the bug introduced? What was the behavior?
If a new feature, why was this architecture chosen? What were the
alternatives?
If a refactor, how is this better than the previous implementation?

If there is a design document for this feature, please link it here.
-->

### Type of Change

<!--
Please check relevant options, delete irrelevant ones.
-->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (non-breaking change that only restructures code)
- [x] Tests (You added tests for code that already exists, or your new
feature included in this PR)
- [ ] Documentation Updates
- [ ] Release

## Before / After

<!--
If just refactoring / back-end changes, this can be just an in-English
description of the change at a technical level.
If a UI change, screenshots should be included.
-->

## Test Plan

<!--
Please describe the tests that you ran to verify your changes and
provide instructions so that others can reproduce.
-->

A new file that tests the `/network/validators` API endpoint is
introduced in this PR.

## Future Tasks
I don't like the return format of `/network/validators` API endpoint. It
appears to cherry-pick most of the fields from the `validators` and
`ballot` table, however some of the fields like `ledger_hash` appear to
have fallen through the cracks.

[This
file](https://github.com/ripple/validator-history-service/blob/51d5db952bb7a776651e7bb67a5fe416195b6d6e/src/api/routes/v1/validator.ts#L109)
contains many examples of hand-maintained user-defined structs.

Is there a current/historical reason for these decision? The code is
much more maintainable if return *all* the columns of a table in the API
endpoint. The end-users are well equipped to filter for useful columns.