Expose WebSocket connection count as prometheus metrics and some refactors (#337)
## High Level Overview of Change

This PR accomplishes the following two tasks:

1. Deprecate `ws_url` and `connected` columns from `crawls` table and
add `connection_health` to capture their true values.
2. Expose number of active WebSocket connections for a network in
prometheus exposition format.

<!--
Please include a summary/list of the changes.
If too broad, please consider splitting into multiple PRs.
If a relevant Asana task, please link it here.
-->

### Context of Change

1. `ws_url` and `connected` columns from `crawls` table were not
capturing all the active WebSocket connections since, we were creating
additional WebSocket connections from using `entry` column of `networks`
table as well.
This change, adds a `connection_health` table that captures all the
WebSocket connections and their status update time. It also, fixes the
bug of `connected` column of `crawls` remaining stale even after loosing
a WebSocket connection.
2. Exposes `/metrics/:network` that can be queried by Grafana to capture
the number of connected nodes for each network and configure alerts.
3. Refactor the code to use xrpl.js library to fetch EnableAmendment
Transaction and Amendments ledger entry as opposed to listening
WebSocket events in `setHandlers` in order to make `setHandlers` more
cleaner.
<!--
Please include the context of a change.
If a bug fix, when was the bug introduced? What was the behavior?
If a new feature, why was this architecture chosen? What were the
alternatives?
If a refactor, how is this better than the previous implementation?

If there is a design document for this feature, please link it here.
-->

### Type of Change

<!--
Please check relevant options, delete irrelevant ones.
-->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] Refactor (non-breaking change that only restructures code)
- [x] Tests (You added tests for code that already exists, or your new
feature included in this PR)
- [x] Documentation Updates
- [ ] Release

## Before / After

<!--
If just refactoring / back-end changes, this can be just an in-English
description of the change at a technical level.
If a UI change, screenshots should be included.
-->

No change in the working on validator-history-service

## Test Plan

<!--
Please describe the tests that you ran to verify your changes and
provide instructions so that others can reproduce.
-->

Added tests in /tests directory to check for /metrics/:network endpoint,
database updates, selects, and utils methods that were introduced by
this change.
I will deploy this branch on staging to check its working on stating
before releasing on production.

<!--
## Future Tasks
For future tasks related to PR.
-->

---------

Co-authored-by: Raj Patel <patelraj22777@gmail.com>