Update responses of API endpoints (#347)
This PR updates the responses of the API endpoints with non-breaking
changes.

- Error Cases are appended with explanatory error messages (commit:
33e141998af530bb9c411505c8e0df0ef05fb8f0). This does not incur any
privacy violations. The VHS software is open-source and can be
deployed/run by anybody in the world. The stack-traces/logs in the
production VHS are identical to any other deployments of the VHS.

- All the API responses are associated with an HTTP status code
(success: 200; failure: 404, 500). This is helpful for monitoring the
health of the service.

The existing unit tests have been updated to account for HTTP status
codes. Mock functions have been used in-place of the status method.