Migrate ESLint to v9 and it's related plugins to their latest major version (#380)
## High Level Overview of Change

This PR has changes similar to
https://github.com/XRPLF/xrpl.js/pull/3075:

1. Update ESLint and its related plugins to their latest major versions.
2. Account for breaking changes introduced by ESLint v9 itself and other
ESLint plugins and configs.
3. Update @xrplf/eslint-config to 2.0.0. See
https://github.com/XRPLF/typescript-style/pull/25
4. Update prettier from v2 to v3.6.2
5. eslint-plugin-jsdoc plugin does not support Node version 18. Updated
CI to use 20.x, 22.x. Production [Docker
file](https://github.com/xpring-eng/vhs-deploy/blob/main/Dockerfile)
uses uses Node 20.


<!--
Please include a summary/list of the changes.
If too broad, please consider splitting into multiple PRs.
If a relevant Asana task, please link it here.
-->

### Context of Change

The plugins and outdated ESLint dependencies from @xrplf/eslint-config
was causing depandabot version update failures and many were piled up.
This PR will resolve them.

1. Reduce number of vulnerabilities

```
Before: 27 vulnerabilities (1 low, 18 moderate, 6 high, 2 critical)
After: 21 vulnerabilities (17 moderate, 4 high)
```
<!--
Please include the context of a change.
If a bug fix, when was the bug introduced? What was the behavior?
If a new feature, why was this architecture chosen? What were the
alternatives?
If a refactor, how is this better than the previous implementation?

If there is a design document for this feature, please link it here.
-->

### Type of Change

<!--
Please check relevant options, delete irrelevant ones.
-->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] Refactor (non-breaking change that only restructures code)
- [ ] Tests (You added tests for code that already exists, or your new
feature included in this PR)
- [ ] Documentation Updates
- [ ] Release

## Before / After

<!--
If just refactoring / back-end changes, this can be just an in-English
description of the change at a technical level.
If a UI change, screenshots should be included.
-->

## Test Plan
Unit, Integration and Linting should pass in CI.
<!--
Please describe the tests that you ran to verify your changes and
provide instructions so that others can reproduce.
-->

<!--
## Future Tasks
For future tasks related to PR.
-->