Mitigate DB deadlock on `manifest` table (#401)
## High Level Overview of Change
feat: introduce retry logic for victim transactions during DB deadlock.
Add unit tests to simulate this scenario

The database deadlock issue has been observed a few times over the last
30 days in the staging and prod environments of the VHS. The VHS
encounters a fatal crash after the occurrence of the deadlock, due to
access of an undefined variable. Both of these issues are mitigated in
this PR. A new exponential transaction retry mechanism has been added.

Unfortunately, we do not have database transaction logs that go beyond
3-4 days time period. So, it is impossible to determine which two
competing database transactions caused this deadlock.

<!--
Please include a summary/list of the changes.
If too broad, please consider splitting into multiple PRs.
If a relevant Asana task, please link it here.
-->


<!--
Please include the context of a change.
If a bug fix, when was the bug introduced? What was the behavior?
If a new feature, why was this architecture chosen? What were the
alternatives?
If a refactor, how is this better than the previous implementation?

If there is a design document for this feature, please link it here.
-->

### Type of Change

<!--
Please check relevant options, delete irrelevant ones.
-->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (non-breaking change that only restructures code)
- [ ] Tests (You added tests for code that already exists, or your new
feature included in this PR)
- [ ] Documentation Updates
- [ ] Release


<!--
If just refactoring / back-end changes, this can be just an in-English
description of the change at a technical level.
If a UI change, screenshots should be included.
-->

## Test Plan
Unit tests that simulate a database deadlock have been added.
<!--
Please describe the tests that you ran to verify your changes and
provide instructions so that others can reproduce.
-->

<!--
## Future Tasks
For future tasks related to PR.
-->