Retain UNL validators in purge (#400)
## High Level Overview of Change

<!--
Please include a summary/list of the changes.
If too broad, please consider splitting into multiple PRs.
If a relevant Asana task, please link it here.
-->
With current logic, if a validator has not participated in the consensus
process for more than a week, it will be removed from the list. This led
to an UNL validator, even though still in the UNL list, got removed by
the logic.
This PR would:
- Retain inactive UNL validators that are still discoverable in the UNL
list in the database, until it got removed from the list.
- Increase the purging period from one week to thirty days.

Resolves https://github.com/ripple/validator-history-service/issues/395
Replaces https://github.com/ripple/validator-history-service/pull/397

### Type of Change

<!--
Please check relevant options, delete irrelevant ones.
-->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (non-breaking change that only restructures code)
- [ ] Tests (You added tests for code that already exists, or your new
feature included in this PR)
- [ ] Documentation Updates
- [ ] Release